requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
fastapi-cli==0.0.5
gensim==4.3.3
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
nltk==3.9.1
numpy==1.26.4
pandas==2.2.3
psycopg2-binary==2.9.10
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
rich==13.9.4
scikit-learn==1.5.2
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
starlette==0.41.3
textblob==0.18.0.post0
threadpoolctl==3.5.0
tqdm==4.67.1
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
wrapt==1.17.0
